# openapi-data-gpt.yaml
openapi: 3.1.0
info:
  title: Data GPT Orchestrator API
  version: "0.1.0"
servers:
  - url: http://localhost:8787
paths:
...
  /orchestrate/run:
    post:
      summary: Run the orchestrator on a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task: { type: string }
      responses:
        '200':
          description: ok
  /web/search:
    get:
      summary: Web search
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: provider
          schema: { type: string, enum: [serpapi, duckduckgo] }
      responses:
        '200':
          description: ok
  /http/get:
    get:
      summary: Fetch page
      parameters:
        - in: query
          name: url
          required: true
          schema: { type: string, format: uri }
      responses:
        '200':
          description: ok
  /llm/gemini/complete:
    post:
      summary: Gemini completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
                model: { type: string }
      responses:
        '200':
          description: ok
...